Change query to ORM for users

+++ internal/models/Post.go
@@ -3,11 +3,11 @@ package models
 import "time"
 
 type Post struct {
-	ID        uint   `gorm:"primaryKey"`
+	ID        int64  `gorm:"primaryKey"`
 	Title     string `gorm:"not null"`
 	Content   string `gorm:"not null"`
 	CreatedAt time.Time
-	UserID    uint
+	UserID    int64
 	User      User
 	Image     *string
 	Likes     []Like
